% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_workflow_set.R
\name{as_workflow_set}
\alias{as_workflow_set}
\title{Save tuning or resampling results as a workflow set}
\usage{
as_workflow_set(object)
}
\arguments{
\item{object}{A named list of results. Names should be unique and the
elements in \code{object} should have at least one of the following classes:
\code{iteration_results}, \code{tune_results}, \code{resample_results}, or \code{tune_race}. Each
element should also contain the original workflow (accomplished using the
\code{save_workflow} option in the control function).}
}
\value{
A workflow set. Note that the \code{option} column will not reflect the
options that were used to create each object.
}
\description{
Save tuning or resampling results as a workflow set
}
