% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes.R
\name{note_add}
\alias{note_add}
\alias{note_get}
\alias{note_reset}
\title{Work with notes in a workflow}
\usage{
note_add(x, id, ..., append = TRUE, collapse = "\\n")

note_get(x, id)

note_reset(x, id)
}
\arguments{
\item{x}{A workflow set}

\item{id}{A single character string for a value in the \code{wflow_id} column.}

\item{...}{One or more character strings.}

\item{append}{A logical value to determine if the new note should be added
to the existing values.}

\item{collapse}{A character string that separates the notes.}
}
\value{
\code{note_add()} and \code{note_reset()} return an updated workflow set.
\code{note_get()} returns a character string.
}
\description{
\code{note_add()} can be used to log important information about the workflow or
its results as you work. Notes can be appended or removed.
}
\examples{
two_class_set \%>\% note_get("none_cart")

new_set <-
   two_class_set \%>\%
   note_add("none_cart", "What does 'cart' stand for\u2753") \%>\%
   note_add("none_cart", "Classification And Regression Trees.")

new_set \%>\% note_get("none_cart") \%>\% cat("\n")

new_set \%>\% note_reset("none_cart") \%>\% note_get("none_cart")
}
