% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{option_list}
\alias{option_list}
\title{Make a classed list of options}
\usage{
option_list(...)
}
\arguments{
\item{...}{A set of named options (or nothing)}
}
\value{
A classed list.
}
\description{
This function returns a named list with an extra class of
\code{"workflow_set_options"} that has corresponding formatting methods for
printing inside of tibbles.
}
\examples{
option_list(a = 1, b = 2)
option_list()
}
