% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pull_workflow_result}
\alias{pull_workflow_result}
\alias{pull_workflow}
\title{Extract a workflow from a workflow set}
\usage{
pull_workflow_result(x, id)

pull_workflow(x, id)
}
\arguments{
\item{x}{A workflow set.}

\item{id}{A single character string for a workflow ID}
}
\value{
The value in the \code{result} or \code{object} field of the workflow set,
depending on the function.
}
\description{
Extract a workflow from a workflow set
}
