% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pull_workflow_set_result}
\alias{pull_workflow_set_result}
\alias{pull_workflow}
\title{Extract elements from a workflow set}
\usage{
pull_workflow_set_result(x, id)

pull_workflow(x, id)
}
\arguments{
\item{x}{A workflow set.}

\item{id}{A single character string for a workflow ID.}
}
\value{
\code{pull_workflow_set_result()} produces a \code{tune_result} or
\code{resample_results} object. \code{pull_workflow()} returns an unfit workflow
object.
}
\description{
\code{pull_workflow_set_result()} retrieves the results of \code{\link[=workflow_map]{workflow_map()}} for a
particular workflow while \code{pull_workflow()} extracts the unfitted workflow
from the \code{info} column.
}
\examples{
library(tune)

pull_workflow_set_result(two_class_res, "none_cart")

pull_workflow(two_class_res, "none_cart")
}
