% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_imports.R, R/fit_best.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{collect_metrics}
\alias{collect_predictions}
\alias{collect_notes}
\alias{collect_extracts}
\alias{\%>\%}
\alias{autoplot}
\alias{extract_spec_parsnip}
\alias{extract_recipe}
\alias{extract_fit_parsnip}
\alias{extract_fit_engine}
\alias{extract_mold}
\alias{extract_preprocessor}
\alias{extract_workflow}
\alias{extract_parameter_set_dials}
\alias{extract_parameter_dials}
\alias{fit_best}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{\%>\%}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}

  \item{hardhat}{\code{\link[hardhat]{extract_fit_engine}}, \code{\link[hardhat]{extract_fit_parsnip}}, \code{\link[hardhat]{extract_mold}}, \code{\link[hardhat]{extract_parameter_dials}}, \code{\link[hardhat]{extract_parameter_set_dials}}, \code{\link[hardhat]{extract_preprocessor}}, \code{\link[hardhat]{extract_recipe}}, \code{\link[hardhat]{extract_spec_parsnip}}, \code{\link[hardhat]{extract_workflow}}}

  \item{tune}{\code{\link[tune]{collect_extracts}}, \code{\link[tune]{collect_metrics}}, \code{\link[tune]{collect_notes}}, \code{\link[tune]{collect_predictions}}, \code{\link[tune]{fit_best}}}
}}

