% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{selectors}
\alias{selectors}
\title{Raw column selectors for modeling}
\usage{
selectors(outcomes, predictors, blueprint = NULL)
}
\arguments{
\item{outcomes, predictors}{Tidyselect expressions specifying the terms of
the model. \code{outcomes} is evaluated first, and then all outcome columns are
removed from the data before \code{predictors} is evaluated. See
\code{\link[workflows:add_variables]{workflows::add_variables()}} for more information. The selectors are not
evaluated until the model is fit.}

\item{blueprint}{A hardhat blueprint used for fine tuning the preprocessing.}
}
\value{
A object with class \code{'selectors'}.
}
\description{
Raw column selectors for modeling
}
\examples{
selectors(mpg, c(cyl, disp))
}
