% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{two_class_set}
\alias{two_class_set}
\alias{two_class_res}
\alias{chi_features_set}
\alias{chi_features_res}
\title{Example Data Sets}
\value{
Workflow sets.
}
\description{
Example Data Sets
}
\details{
Example workflow sets and associated model fits.

\code{two_class_set} and \code{two_class_res} were generated using the data in the
package file \code{example-data/two-class-set.R}

\code{chi_features_set} and \code{chi_features_res} were generated using the data in the
package file \code{example-data/chi-features-res.R}. It is meant to approximate
the sequence of models built in Section 1.3 of Kuhn and Johnson (2019).
}
\examples{
data(two_class_set)
two_class_set
}
\references{
Max Kuhn and Kjell Johnson (2019) \emph{Feature Engineering and
Selection}, \url{https://bookdown.org/max/FES/a-more-complex-example.html}
}
\keyword{datasets}
