% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_workflow_model}
\alias{update_workflow_model}
\title{Update workflow componets inside a workflow set}
\usage{
update_workflow_model(x, id, spec, formula = NULL)
}
\arguments{
\item{x}{A workflow set.}

\item{id}{A character string for values from the \code{wflow_id} column that
indicates which workflow to update.}

\item{spec}{A parsnip model specification.}

\item{formula}{An optional formula override to specify the terms of the
model. Typically, the terms are extracted from the formula or recipe
preprocessing methods. However, some models (like survival and bayesian
models) use the formula not to preprocess, but to specify the structure of
the model. In those cases, a formula specifying the model structure must be
passed unchanged into the model call itself. This argument is used for those
purposes.}

\item{recipe}{A recipe created using \code{\link[recipes:recipe]{recipes::recipe()}}.}

\item{blueprint}{A hardhat blueprint used for fine tuning the preprocessing.
If \code{NULL}, \code{\link[hardhat:default_recipe_blueprint]{hardhat::default_recipe_blueprint()}} is used. Note that
preprocessing done here is separate from preprocessing that might be done
automatically by the underlying model.}
}
\value{
An altered workflow set where the \code{results} column value for the
selected  \code{id} have been reset to empty.
}
\description{
Update workflow componets inside a workflow set
}
