% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_imports.R
\docType{package}
\name{workflowsets-package}
\alias{workflowsets}
\alias{workflowsets-package}
\title{workflowsets: Create a Collection of 'tidymodels' Workflows}
\description{
A workflow is a combination of a model and preprocessors (e.g, a formula, recipe, etc.) (Kuhn and Silge (2021) \url{https://www.tmwr.org/}). In order to try different combinations of these, an object can be created that contains many workflows. There are functions to create workflows en masse as well as training them and visualizing the results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidymodels/workflowsets}
  \item \url{https://workflowsets.tidymodels.org}
  \item Report bugs at \url{https://github.com/tidymodels/workflowsets/issues}
}

}
\author{
\strong{Maintainer}: Hannah Frick \email{hannah@posit.co} (\href{https://orcid.org/0000-0002-6049-5258}{ORCID})

Authors:
\itemize{
  \item Max Kuhn \email{max@posit.co} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})
  \item Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})
}

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
