% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_workflow_set.R
\name{as_workflow_set}
\alias{as_workflow_set}
\title{Save results from tuning or resampling functions as a workflow set}
\usage{
as_workflow_set(...)
}
\arguments{
\item{...}{One or more named objects. Names should be unique and the
objects should have at least one of the following classes:
\code{iteration_results}, \code{tune_results}, \code{resample_results}, or \code{tune_race}. Each
element should also contain the original workflow (accomplished using the
\code{save_workflow} option in the control function).}
}
\value{
A workflow set. Note that the \code{option} column will not reflect the
options that were used to create each object.
}
\description{
If results have been generated directly from functions like
\code{\link[tune:tune_grid]{tune::tune_grid()}}, they can be combined into a workflow set using this
function.
}
