% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.workflow_set}
\alias{autoplot.workflow_set}
\title{Plot the results of a workflow set}
\usage{
\method{autoplot}{workflow_set}(
  object,
  rank_metric = NULL,
  metric = NULL,
  id = "workflow_set",
  select_best = FALSE,
  std_errs = qnorm(0.95),
  type = "class",
  ...
)
}
\arguments{
\item{object}{A \code{workflow_set} whose elements have results.}

\item{rank_metric}{A character string for which metric should be used to rank
the results. If none is given, the first metric in the metric set is used
(after filtering by the \code{metric} option).}

\item{metric}{A character vector for which metrics (apart from \code{rank_metric})
to be included in the visualization.}

\item{id}{A character string for what to plot. If a value of
\code{"workflow_set"} is used, the results of each model (and sub-model) are ordered
and plotted. Alternatively, a value of the workflow set's \code{wflow_id} can be
given and the \code{autoplot()} method is executed on that workflow's results.}

\item{select_best}{A logical; should the results only contain the numerically
best submodel per workflow?}

\item{std_errs}{The number of standard errors to plot (if the standard error
exists).}

\item{type}{The aesthetics with which to differentiate workflows. The
default \code{"class"} maps color to the model type and shape to the preprocessor
type. The \code{"workflow"} option maps a color to each \code{"wflow_id"}. This
argument is ignored for values of \code{id} other than \code{"workflow_set"}.}

\item{...}{Other options to pass to \code{autoplot()}.}
}
\value{
A ggplot object.
}
\description{
This \code{autoplot()} method plots performance metrics that have been ranked using
a metric. It can also run \code{autoplot()} on the individual results (per
\code{wflow_id}).
}
\details{
This function is intended to produce a default plot to visualize helpful
information across all possible applications of a workflow set. A more
appropriate plot for your specific analysis can be created by
calling \code{\link[=rank_results]{rank_results()}} and using standard \code{ggplot2} code for plotting.

The x-axis is the workflow rank in the set (a value of one being the best)
versus the performance metric(s) on the y-axis. With multiple metrics, there
will be facets for each metric.

If multiple resamples are used, confidence bounds are shown for each result
(90\% confidence, by default).
}
\note{
The package supplies two pre-generated workflow sets, \code{two_class_set}
and \code{chi_features_set}, and associated sets of model fits
\code{two_class_res} and \code{chi_features_res}.

The \verb{two_class_*} objects are based on a binary classification problem
using the \code{two_class_dat} data from the modeldata package. The six
models utilize either a bare formula or a basic recipe utilizing
\code{recipes::step_YeoJohnson()} as a preprocessor, and a decision tree,
logistic regression, or MARS model specification. See \code{?two_class_set}
for source code.

The \verb{chi_features_*} objects are based on a regression problem using the
\code{Chicago} data from the modeldata package. Each of the three models
utilize a linear regression model specification, with three different
recipes of varying complexity. The objects are meant to approximate the
sequence of models built in Section 1.3 of Kuhn and Johnson (2019). See
\code{?chi_features_set} for source code.
}
\examples{
autoplot(two_class_res)
autoplot(two_class_res, select_best = TRUE)
autoplot(two_class_res, id = "yj_trans_cart", metric = "roc_auc")
}
