% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_metrics.workflow_set}
\alias{collect_metrics.workflow_set}
\title{Obtain and format results produced by tuning functions for workflow sets}
\usage{
\method{collect_metrics}{workflow_set}(x, summarize = TRUE, ...)
}
\arguments{
\item{x}{A \code{workflow_set} object where all workflows have been evaluated.}

\item{summarize}{A logical for whether the performance estimates should be
summarized via the mean (over resamples) or the raw performance values (per
resample) should be returned along with the resampling identifiers.}

\item{...}{Not currently used.}
}
\value{
A tibble.
}
\description{
Obtain and format results produced by tuning functions for workflow sets
}
