% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_workflow_set_result}
\alias{extract_workflow_set_result}
\alias{extract_workflow.workflow_set}
\alias{extract_spec_parsnip.workflow_set}
\alias{extract_recipe.workflow_set}
\alias{extract_fit_parsnip.workflow_set}
\alias{extract_fit_engine.workflow_set}
\alias{extract_mold.workflow_set}
\alias{extract_preprocessor.workflow_set}
\title{Extract elements from a workflow set}
\usage{
extract_workflow_set_result(x, id, ...)

\method{extract_workflow}{workflow_set}(x, id, ...)

\method{extract_spec_parsnip}{workflow_set}(x, id, ...)

\method{extract_recipe}{workflow_set}(x, id, ...)

\method{extract_fit_parsnip}{workflow_set}(x, id, ...)

\method{extract_fit_engine}{workflow_set}(x, id, ...)

\method{extract_mold}{workflow_set}(x, id, ...)

\method{extract_preprocessor}{workflow_set}(x, id, ...)
}
\arguments{
\item{x}{A workflow set.}

\item{id}{A single character string for a workflow ID.}

\item{...}{Other options (not currently used).}
}
\value{
\code{extract_workflow_set_result()} produces a \code{tune_result} or
\code{resample_results} object. \code{extract_workflow()} returns an unfit workflow
object.
}
\description{
\code{extract_workflow_set_result()} retrieves the results of \code{\link[=workflow_map]{workflow_map()}} for a
particular workflow while \code{extract_workflow()} extracts the unfitted workflow
from the \code{info} column. These functions supersede \code{pull_workflow_set_result()}
and \code{pull_workflow()} (since their names are more consistent with other
tidymodels functions).
}
\examples{
library(tune)

extract_workflow_set_result(two_class_res, "none_cart")

extract_workflow(two_class_res, "none_cart")
}
