% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tbl_sum}
\alias{type_sum}
\alias{size_sum}
\alias{collect_metrics}
\alias{collect_predictions}
\alias{\%>\%}
\alias{autoplot}
\alias{extract_spec_parsnip}
\alias{extract_recipe}
\alias{extract_fit_parsnip}
\alias{extract_fit_engine}
\alias{extract_mold}
\alias{extract_preprocessor}
\alias{extract_workflow}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}

  \item{hardhat}{\code{\link[hardhat:hardhat-extract]{extract_fit_engine}}, \code{\link[hardhat:hardhat-extract]{extract_fit_parsnip}}, \code{\link[hardhat:hardhat-extract]{extract_mold}}, \code{\link[hardhat:hardhat-extract]{extract_preprocessor}}, \code{\link[hardhat:hardhat-extract]{extract_recipe}}, \code{\link[hardhat:hardhat-extract]{extract_spec_parsnip}}, \code{\link[hardhat:hardhat-extract]{extract_workflow}}}

  \item{tibble}{\code{\link[tibble:reexports]{size_sum}}, \code{\link[tibble:reexports]{tbl_sum}}, \code{\link[tibble:reexports]{type_sum}}}

  \item{tune}{\code{\link[tune:collect_predictions]{collect_metrics}}, \code{\link[tune]{collect_predictions}}}
}}

