% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbd.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tune_grid}
\alias{tune_bayes}
\alias{fit_resamples}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}}

  \item{tune}{\code{\link[tune]{fit_resamples}}, \code{\link[tune]{tune_bayes}}, \code{\link[tune]{tune_grid}}}
}}

